pyproject.toml
[tool.poetry]
name = "alchemy"
version = "0.1.0"
description = ""
authors = ["Shaun Martin <inhumantsar@protonmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
ipykernel = "^6.22.0"
llama-index = "^0.5.9"
nest-asyncio = "^1.5.6"
python-dotenv = "^1.0.0"
black = "^23.3.0"
pytest = "^7.2.2"
nltk = "^3.8.1"
notebook = "^6.5.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 140